requirements.txt
absl-py==2.1.0
amqp==5.3.1
astunparse==1.6.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Flask==3.1.0
flatbuffers==24.12.23
flower==2.0.1
flwr==1.14.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.3
h5py==3.12.1
humanize==4.11.0
idna==3.10
iterators==0.0.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kombu==5.4.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==4.25.5
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
tomli_w==1.1.0
tornado==6.4.2
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.1
